requirements.txt
pyyaml
numpy==1.22.0
scipy==1.7.3
numba==0.58.1
cycler
ipykernel
kiwisolver
tqdm 
rich 
pyglet==1.5.27
psutil
setuptools==65.5.0
wheel==0.38.4
wandb==0.15.12
tensorboard
cairosvg
PyPDF2
gymnasium
moviepy
seaborn